image_processing/requirements.txt
Pillow==8.4.0
numpy==1.21.0
requests==2.26.0
matplotlib==3.4.3
